Move all the interfaces to a single file (#185)
