--- src/scenes/Dashboard/scenes/ManageMentees/interfaces.tsx
@@ -1,9 +1,0 @@
-import { Profile, Mentor } from '../../../../interfaces';
-
-export interface Mentee {
-  id: number;
-  profile: Profile;
-  state: string;
-  submissionUrl: string;
-  mentor: Mentor;
-}
