+++ src/scenes/Dashboard/scenes/ManageMentors/components/MentorRow/interfaces.ts
@@ -1,18 +1,6 @@
-import { Mentor, Question } from '../../../../../../interfaces';
+import { Mentor } from '../../../../../../types';
 
 export interface Props {
   mentor: Mentor;
   programState: string;
-}
-
-export interface MentorResponse {
-  id: ID;
-  question: Question;
-  mentor: Mentor;
-  response: string;
-}
-
-export interface ID {
-  questionId: number;
-  mentorId: number;
 }
