--- src/scenes/Dashboard/scenes/ManageMentors/interfaces.tsx
@@ -1,12 +1,0 @@
-export interface Mentor {
-  id: number;
-  profile: {
-    firstName: string,
-    lastName: string,
-    imageUrl: string,
-    linkedinUrl: string,
-    headline: string,
-  };
-  state: string;
-  prerequisites: string;
-}
