+++ src/scenes/Home/components/MentorPrograms/components/MentorProgramCard/interfaces.ts
@@ -1,4 +1,4 @@
-import { SavedProgram } from '../../../../../../interfaces';
+import { SavedProgram } from '../../../../../../types';
 
 export interface MentorProgramCardProps {
   program: SavedProgram;
