+++ src/scenes/Home/components/MentorPrograms/index.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from 'react';
 import { Button, Col, notification, Result, Row, Spin, Typography } from 'antd';
 import styles from '../../styles.css';
 import axios, { AxiosResponse } from 'axios';
-import { SavedProgram } from '../../../../interfaces';
+import { SavedProgram } from '../../../../types';
 import { SmileOutlined } from '@ant-design/icons';
 import { API_URL } from '../../../../constants';
 import MentorProgramCard from './components/MentorProgramCard';
