+++ src/scenes/Home/components/NavigationBar/interfaces.ts
@@ -1,4 +1,4 @@
-import { Profile } from '../../interfaces';
+import { Profile } from '../../../../types';
 
 export interface NavigationBarProps {
   user?: Profile | null;
