+++ src/scenes/Home/scenes/RequestMentors/components/MentorCard/interfaces.ts
@@ -1,4 +1,4 @@
-import { Mentor } from '../../../../../../interfaces';
+import { Mentor } from '../../../../../../types';
 
 export interface MentorCardProps {
   item: Mentor;
